// openpectus/frontend/src/app/api/models/MethodAndState.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Method } from './Method';
import type { MethodState } from './MethodState';

export type MethodAndState = {
    method: Method;
    state: MethodState;
};